ai-service/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
python-multipart
requests
numpy
pillow
pyyaml
